+++ src/Shared/Types/Subreddit.fs
@@ -1,3 +1,3 @@
 namespace RIDOTypes
 
-type Subreddit = { subreddit: string; favourited: bool; lastUpdated: int }
+type Subreddit = { subreddit: string; favourited: bool; lastUpdated: int64 }
